refactor(codellm): P2 extract code-execution + sandbox from agentruntime into internal/coderun
