+++ .golangci.yaml
@@ -487,7 +487,7 @@ linters:
       - path: 'internal/gitapi/'
         linters:
           - gosec  # Git operations require subprocess calls which are legitimate
-      - path: 'internal/agentruntime/coderun\.go'
+      - path: 'internal/coderun/coderun\.go'
         linters:
           - gosec  # code runner legitimately execs operator-allowlisted interpreters
       - path: 'internal/mdloader/highlight/'
