+++ internal/coderun/coderun_debug_test.go
@@ -1,4 +1,4 @@
-package agentruntime
+package coderun
 
 import (
 	"context"
