+++ internal/coderun/coderun_main_test.go
@@ -1,4 +1,4 @@
-package agentruntime
+package coderun
 
 import (
 	"context"
